+++ samples/msal-node-samples/auth-code/test/auth-code-aad.spec.ts
@@ -173,7 +173,10 @@ describe("Auth Code AAD PPE Tests", () => {
             await page.goto(`${homeRoute}/?prompt=login&loginHint=${USERNAME}`, {waitUntil: "networkidle0"});
             await page.waitForSelector("#i0116");
             const emailInput = await page.$("#i0116");
-            const email = await page.evaluate(element => element.value, emailInput);
+            const email = await page.evaluate(element => {
+                const emailInput = element as HTMLInputElement;
+                return emailInput.value;
+            }, emailInput);
             expect(email).toBe(USERNAME);
         });